requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
google-api-python-client
werkzeug
Flask-Session
pymysql
cryptography
gunicorn